# lab/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for calculating ion concentration in electrolyte mixtures.",
        "title": "Electrolyte Concentration API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/electrolytes": {
            "get": {
                "description": "Returns all electrolytes (public access)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Electrolytes"
                ],
                "summary": "Get list of electrolytes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: []Electrolyte",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/electrolytes/{id}": {
            "get": {
                "description": "Returns detailed information about a specific electrolyte",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Electrolytes"
                ],
                "summary": "Get electrolyte by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Electrolyte ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: Electrolyte",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}